Split HIR type lowering helpers (#1299)
Move syntax-to-HIR literal, type, and operator lowering out of hir.rs into hir/types.rs while preserving the internal compiler.hir API. Update the monolith reporter and decomposition ratchets for the lower HIR facade line count.

Refs #1254.
