+++ docs/compiler-source-decomposition-plan.md
@@ -45,7 +45,9 @@ count. It also raised the top-seven share because test code left
 analysis extraction then split generic inference and monomorphization into a
 tracked `compiler.hir` module. The HIR model extraction then moved public HIR
 types and type-display helpers into `stage1/crates/axiomc/src/hir/model.rs`,
-lowering both absolute top-file lines and share.
+and the HIR type-lowering extraction moved syntax-to-HIR literal, type, and
+operator lowering into `stage1/crates/axiomc/src/hir/types.rs`, lowering both
+absolute top-file lines and share.
 
 ## Current Top Files
 
@@ -54,7 +56,7 @@ Snapshot from 2026-07-02:
 | Rank | Current Rust file | Lines | Target package boundary | First extraction slice |
 | ---: | --- | ---: | --- | --- |
 | 1 | `stage1/crates/axiomc/src/cranelift_backend.rs` | 27,994 | `compiler.backend.native` | Split direct-native lowering by runtime ABI groups: scalar/aggregate value features, capability shims, host imports, object emission, unsupported diagnostics, and evidence helpers. |
-| 2 | `stage1/crates/axiomc/src/hir.rs` | 12,117 | `compiler.hir` | Generic inference and monomorphization now live in `stage1/crates/axiomc/src/hir/generics.rs`; public HIR model types now live in `stage1/crates/axiomc/src/hir/model.rs`; next split name resolution, type checking, capability analysis, ownership/borrow validation, property clauses, and HIR diagnostics behind the package APIs in `docs/compiler-hir-ownership-capability.md`. |
+| 2 | `stage1/crates/axiomc/src/hir.rs` | 11,885 | `compiler.hir` | Generic inference and monomorphization now live in `stage1/crates/axiomc/src/hir/generics.rs`; public HIR model types now live in `stage1/crates/axiomc/src/hir/model.rs`; syntax-to-HIR type/literal lowering now lives in `stage1/crates/axiomc/src/hir/types.rs`; next split name resolution, type checking, capability analysis, ownership/borrow validation, property clauses, and HIR diagnostics behind the package APIs in `docs/compiler-hir-ownership-capability.md`. |
 | 3 | `stage1/crates/axiomc/src/project.rs` | 10,812 | `compiler.package_graph`, `compiler.commands`, `compiler.evidence` | Split manifest/workspace loading, command orchestration, provenance/debug records, and build artifact planning along package ownership. |
 | 4 | `stage1/crates/axiomc/src/main.rs` | 10,678 | `compiler.commands` | Move command parsing, JSON envelope construction, check/build/run/test/doc/trace orchestration, and exit handling behind `docs/compiler-command-lsp-packages.md` APIs. |
 | 5 | `stage1/crates/axiomc/src/codegen.rs` | 7,804 | `compiler.backend.generated_rust`, `compiler.backend.contracts` | Isolate generated-Rust compatibility emission from backend target selection and unsupported-feature contracts. |
@@ -71,16 +73,17 @@ matching ceiling in this table in the same PR.
 
 | Tracked item | Ceiling |
 | --- | ---: |
-| `summary.top_file_line_share` | 0.8979 |
-| `summary.top_file_lines` | 79934 |
+| `summary.top_file_line_share` | 0.8952 |
+| `summary.top_file_lines` | 79702 |
 | `stage1/crates/axiomc/src/cranelift_backend.rs` | 27994 |
-| `stage1/crates/axiomc/src/hir.rs` | 12117 |
+| `stage1/crates/axiomc/src/hir.rs` | 11885 |
 | `stage1/crates/axiomc/src/project.rs` | 10812 |
 | `stage1/crates/axiomc/src/main.rs` | 10678 |
 | `stage1/crates/axiomc/src/codegen.rs` | 7804 |
 | `stage1/crates/axiomc/src/syntax.rs` | 6324 |
 | `stage1/crates/axiomc/src/hir/generics.rs` | 4205 |
 | `stage1/crates/axiomc/src/hir/model.rs` | 607 |
+| `stage1/crates/axiomc/src/hir/types.rs` | 241 |
 | `stage1/crates/axiomc/src/registry.rs` | 2159 |
 | `stage1/crates/axiomc/src/lib.rs` | 21 |
 
@@ -92,9 +95,10 @@ matching ceiling in this table in the same PR.
 2. `compiler.backend.contracts`: move target selection and unsupported-feature
    contracts out of generated-Rust code before the final generated-Rust removal
    gate.
-3. `compiler.hir`: generic inference/monomorphization and public HIR model
-   types are split; continue with resolution, typing, capability, ownership,
-   and property checks in that order so diagnostics stay stable.
+3. `compiler.hir`: generic inference/monomorphization, public HIR model types,
+   and syntax-to-HIR type/literal lowering are split; continue with resolution,
+   typing, capability, ownership, and property checks in that order so
+   diagnostics stay stable.
 4. `compiler.commands` and `compiler.package_graph`: separate command envelopes
    from package loading so the snapshot bootstrap can invoke package APIs
    without Cargo assumptions.
